move store stuff into com.socrata
