+++ store-redshift/src/main/scala/com/socrata/config/RedshiftSecondaryConfig.java
@@ -1,9 +1,7 @@
-package config;
+package com.socrata.config;
 
 
 import com.socrata.datacoordinator.secondary.SecondaryWatcherAppConfig;
-import io.quarkus.runtime.annotations.StaticInitSafe;
-import io.smallrye.config.ConfigMapping;
 
 public interface RedshiftSecondaryConfig extends SecondaryWatcherAppConfig {
