+++ store-redshift/src/main/scala/com/socrata/config/ZookeeperConfig.java
@@ -1,4 +1,4 @@
-package config;
+package com.socrata.config;
 
 import io.quarkus.runtime.annotations.StaticInitSafe;
 import io.smallrye.config.ConfigMapping;
