+++ store-redshift/src/main/scala/com/socrata/db/meta/entity/DatasetInternalName.scala
@@ -1,4 +1,4 @@
-package db.meta.entity
+package com.socrata.db.meta.entity
 
 import io.quarkus.hibernate.orm.panache.PanacheEntity
 import jakarta.persistence.Entity
