+++ store-redshift/src/main/scala/com/socrata/db/meta/package-info.java
@@ -1,5 +1,5 @@
-@PersistenceUnit("store")
-package db.store;
+@PersistenceUnit("db/meta")
+package com.socrata.db.meta;
 
 import io.quarkus.hibernate.orm.PersistenceUnit;
