+++ store-redshift/src/main/scala/com/socrata/db/meta/service/DatasetService.scala
@@ -1,6 +1,6 @@
-package db.meta.service
+package com.socrata.db.meta.service
 
-import db.meta.repository.DatasetRepository
+import com.socrata.db.meta.repository.DatasetRepository
 import jakarta.enterprise.context.ApplicationScoped
 import jakarta.transaction.Transactional
